Move the boolean tests into the Module tests.
